More changes to mitigate narrow_phase crash
Extracted plant despawning into its own system.
Reorganised system ordering.
