+++ src/lib.rs
@@ -1,7 +1,9 @@
+use behaviour::lifecycle;
 use bevy::{
     prelude::{App, CoreStage, Plugin, SystemSet, SystemStage},
     time::FixedTimestep,
 };
+use bevy_rapier2d::prelude::PhysicsStages;
 
 mod attributes;
 mod behaviour;
@@ -26,6 +28,12 @@ pub fn time_step_system_set() -> SystemSet {
         .with_system(spawning::update_plant_size)
 }
 
+pub fn despawn_system_set() -> SystemSet {
+    SystemSet::new()
+        .with_system(lifecycle::kill_bug_system)
+        .with_system(lifecycle::hatch_egg_system)
+        .with_system(spawning::despawn_plants_system)
+}
 pub struct GenesisPlugin;
 
 impl Plugin for GenesisPlugin {
@@ -38,33 +46,37 @@ impl Plugin for GenesisPlugin {
 
         let spawners = spawning::Spawners::from_configs(&config_instance.spawners).unwrap();
 
-        app.add_stage_after(CoreStage::Update, COLLISIONS, SystemStage::parallel())
-            .add_stage_after(COLLISIONS, CLEAN_UP, SystemStage::parallel())
-            .insert_resource(config::BACKGROUND)
-            .insert_resource(ui::AverageAttributeStatistics::default())
-            .insert_resource(ui::CountStatistics::default())
-            .insert_resource(ui::BugPerformanceStatistics::default())
-            .insert_resource(ui::EnergyStatistics::default())
-            .insert_resource(ui::EntityPanelState::default())
-            .insert_resource(ui::GlobalPanelState::default())
-            .add_startup_system_set(setup::setup_system_set())
-            .add_system_set(ui::interaction_system_set())
-            .add_system_set(ui::selection_system_set())
-            .add_system_set(behaviour::time_step_system_set())
-            .add_system_set(behaviour::egg_spawning_system_set())
-            .add_system_set(behaviour::slow_behaviour_system_set())
-            .add_system_set(behaviour::metabolism_system_set())
-            .add_system_set(ui::global_statistics_system_set())
-            .add_system_set(ui::regular_saver_system_set())
-            .add_system_set(slow_system_set())
-            .add_system_set(time_step_system_set())
-            .add_system_to_stage(CoreStage::Last, ui::save_on_close)
-            .add_system_set_to_stage(COLLISIONS, behaviour::eating_system_set())
-            .add_system_set_to_stage(CLEAN_UP, behaviour::despawn_system_set())
-            .insert_resource(ecosystem::Ecosystem::new(config_instance.world_energy))
-            .insert_resource(spawners)
-            .insert_resource(spawning::PlantSizeRandomiser::new(
-                config_instance.plant_size_range,
-            ));
+        app.add_stage_after(
+            PhysicsStages::Writeback,
+            COLLISIONS,
+            SystemStage::parallel(),
+        )
+        .add_stage_after(COLLISIONS, CLEAN_UP, SystemStage::parallel())
+        .insert_resource(config::BACKGROUND)
+        .insert_resource(ui::AverageAttributeStatistics::default())
+        .insert_resource(ui::CountStatistics::default())
+        .insert_resource(ui::BugPerformanceStatistics::default())
+        .insert_resource(ui::EnergyStatistics::default())
+        .insert_resource(ui::EntityPanelState::default())
+        .insert_resource(ui::GlobalPanelState::default())
+        .insert_resource(ecosystem::Ecosystem::new(config_instance.world_energy))
+        .insert_resource(spawners)
+        .insert_resource(spawning::PlantSizeRandomiser::new(
+            config_instance.plant_size_range,
+        ))
+        .add_startup_system_set(setup::setup_system_set())
+        .add_system_set(ui::interaction_system_set())
+        .add_system_set(ui::selection_system_set())
+        .add_system_set(behaviour::time_step_system_set())
+        .add_system_set(behaviour::egg_spawning_system_set())
+        .add_system_set(behaviour::slow_behaviour_system_set())
+        .add_system_set(behaviour::metabolism_system_set())
+        .add_system_set(ui::global_statistics_system_set())
+        .add_system_set(ui::regular_saver_system_set())
+        .add_system_set(slow_system_set())
+        .add_system_set(time_step_system_set())
+        .add_system_to_stage(CoreStage::Last, ui::save_on_close)
+        .add_system_set_to_stage(COLLISIONS, behaviour::eating_system_set())
+        .add_system_set_to_stage(CLEAN_UP, despawn_system_set());
     }
 }
